Move test helpers to test_helper.lua
